+++ packages/client/src/tx/tx.ts
@@ -1,7 +1,7 @@
 import { ValueCompat } from "@/compatibility"
 import { getCallData } from "@/utils/get-call-data"
 import type { ChainHead$ } from "@polkadot-api/observable-client"
-import { TxCreator, TxCreatorBindings } from "@polkadot-api/polkadot-signer"
+import { TxCreator, TxCreatorBindings } from "@polkadot-api/tx-creator"
 import {
   _void,
   compact,
@@ -92,7 +92,7 @@ export const createTxEntry = <
     )
 
   return (arg?: Arg): Transaction<EC> => {
-    const create$ = <T extends TxCreator<any>>(
+    const create$ = <T extends TxCreator>(
       creator: T,
       opts: TxCreatorOptions<T, EC> | undefined,
       mockedSignature: boolean,
@@ -153,7 +153,7 @@ export const createTxEntry = <
         mergeMap((tx) => submit$(chainHead, broadcast, tx, true)),
       )
 
-    const getPaymentInfoWithOptions = async <T extends TxCreator<any>>(
+    const getPaymentInfoWithOptions = async <T extends TxCreator>(
       creator: T,
       options: TxCreatorOptions<T, EC> | undefined,
     ) => {