+++ packages/signers/raw-tx-creator/src/from-raw-signer.ts
@@ -3,8 +3,9 @@ import {
   TxArgSpec,
   TxCreator,
   TxCreatorEnhancer,
-} from "@polkadot-api/polkadot-signer"
+} from "@polkadot-api/tx-creator"
 import {
+  CommonEnhancersSpecs,
   createV4Tx,
   getSignBytes,
   withCommonExtensions,
@@ -27,7 +28,7 @@ export const getTxCreator = (
   publicKey: Uint8Array,
   signingType: "Ecdsa" | "Ed25519" | "Sr25519",
   sign: (input: Uint8Array) => Promise<Uint8Array> | Uint8Array,
-) => {
+): RawTxCreator => {
   const creator: TxCreator<[]> = async (
     payload,
     _,
@@ -73,6 +74,10 @@ export const getTxCreator = (
     signBytes: getSignBytes(sign),
   })
 }
+export type RawTxCreator = {
+  publicKey: Uint8Array
+  signBytes: (data: Uint8Array) => Promise<Uint8Array>
+} & TxCreator<CommonEnhancersSpecs>
 
 const accId = AccountId().enc
 const SIG = new Uint8Array(65).fill(0xcd)